+++ addons/binding/org.openhab.binding.ebus/src/main/java/org/openhab/binding/ebus/internal/EBusAdvancedLogging.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2010-2017 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.openhab.binding.ebus.internal;
+
+import java.io.File;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import de.csdev.ebus.utils.EBusTelegramWriter;
+
+/**
+ *
+ * @author Christian Sowada - Initial contribution
+ */
+public class EBusAdvancedLogging extends EBusTelegramWriter {
+
+    @SuppressWarnings("unused")
+    private final Logger logger = LoggerFactory.getLogger(EBusAdvancedLogging.class);
+
+    public EBusAdvancedLogging() {
+        super(new File(System.getProperty("openhab.logdir")));
+    }
+
+}