+++ addons/binding/org.openhab.binding.ebus/src/main/java/org/openhab/binding/ebus/internal/EBusLibClient.java
@@ -250,11 +250,6 @@ public class EBusLibClient {
      *
      */
     public void stopClient() {
-        //
-        // if (mockupSynJob != null) {
-        // mockupSynJob.cancel(true);
-        // mockupSynJob = null;
-        // }
 
         if (controller != null && !controller.isInterrupted()) {
             controller.interrupt();